Accelerator paths for AccelGroupEntry.
The accelerator paths for the class AccelGroupEntry should be set based on the
accelerator key and modifier using the Ruby method
Gtk3::AccelGroup.accelerator_name().

Supposedly the accelerator path should be set in a GQuark when querying an
accelerator group. However, this GQuark seems to be set to 0 regardless of the
values that I use.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>